+++ ext/duckdb/table_function.h
@@ -10,6 +10,7 @@ struct _rubyDuckDBTableFunction {
 
 typedef struct _rubyDuckDBTableFunction rubyDuckDBTableFunction;
 
+extern VALUE cDuckDBTableFunction;
 rubyDuckDBTableFunction *get_struct_table_function(VALUE self);
 void rbduckdb_init_duckdb_table_function(void);