+++ test/duckdb_test/table_function_bind_info_test.rb
@@ -3,7 +3,7 @@
 require 'test_helper'
 
 module DuckDBTest
-  class BindInfoTest < Minitest::Test
+  class TableFunctionBindInfoTest < Minitest::Test
     def setup
       @db = DuckDB::Database.open
       @conn = @db.connect
@@ -133,6 +133,19 @@ module DuckDBTest
       end
 
       assert_equal table_function, result
+    end
+
+    def test_bind_info_alias
+      assert_same DuckDB::TableFunction::BindInfo, DuckDB::BindInfo
+    end
+
+    def test_bind_info_alias_deprecation_warning
+      DuckDB.send(:remove_const, :BindInfo) if DuckDB.const_defined?(:BindInfo, false)
+      warning = capture_io { DuckDB::BindInfo }.last
+
+      assert_match(/deprecated/, warning)
+    ensure
+      DuckDB.send(:remove_const, :BindInfo) if DuckDB.const_defined?(:BindInfo, false)
     end
   end
 end